Don't output table if no violations are specified
Right now when using `fail` just to fail the build, but not to output
any information (i.e. just using `fail()`), Danger outputs an empty
table with a message of "undefined", which is annoying. (see #109 for
context)

I _think_ this fixes it, but I'm not sure if it's the right fix and/or
if there's a better way to do it. Please let me know!