Add back space for messages containing markdown.
For github to render markdown in tables there has to be some space
around the message. This will add a <p> and render the markdown.
However, it also leads to some funky vertical spacing due to the nature
of some browser styling for the <p> tag.

This commit looks at the message to determine if it contains markdown,
if yes, then add the space for proper rendering. If not, render the
compacter table to get nice vertical alignment.