[GitLab] Calculate inline comment positions correctly
To be able to correctly add inline comments on GitLab we need to provide the file path pre & post change, along with the line number pre & post change.

Introduced an inlinePositionParser that calculates these values by inspecting the structured diff.

A binary search approach was used to make sure that the work to calculate the changes does not grow exponentially with the number of changes in the diff.