[GitLab] Move to gitbeaker/rest
Major version 36 of gitbeaker contains changes to the type system and the structure of the framework.

Moved to gitbeaker/rest which is the expected package to consume going forward, and adjusted our usage of the types to match the changes in the framework.

Additionally gitbeaker now leverages native-fetch  under the hood which nock does not support. Version 4.0.0 of nock in beta can handle testing native-fetch, but does not yet support recording fixtures. To continue using nock for testing without updating to the beta-version, a fetch polyfill has been introduced to use for the gitlab api tests.